requirements.txt
altair==4.2.0
appdirs==1.4.4
astor==0.8.1
attrs==22.1.0
base58==2.1.1
beautifulsoup4==4.11.1
blinker==1.5
boto3==1.26.12
botocore==1.29.12
bs4==0.0.1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cmake==3.25.0
cssselect==1.2.0
decorator==5.1.1
entrypoints==0.4
enum-compat==0.0.3
fake-useragent==0.1.14
feedparser==6.0.10
idna==3.4
importlib-metadata==5.0.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.0
lxml==4.9.1
MarkupSafe==2.1.1
multitasking==0.0.11
numpy==1.23.4
packaging==21.3
pandas==1.5.1
parse==1.19.0
Pillow==9.3.0
plotly==5.11.0
protobuf==3.20.1
pydeck==0.8.0
pyee==8.2.2
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==1.4.3
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-html==0.10.0
s3transfer==0.6.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.2.post1
streamlit==0.62.0
tenacity==8.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
w3lib==2.0.1
watchdog==2.1.9
websockets==10.4
yahoo-fin==0.8.9.1
yfinance==0.1.87
zipp==3.10.0
